commit loader loading on homepage
