+++ src/components/Accordion.css
@@ -1,4 +1,4 @@
-.accordion {
+/* .accordion {
   max-width: 100%;
   margin: 0px auto 0;
   border-radius: 3px;
@@ -59,7 +59,7 @@
 
 /*****  FORM CSS   *****/
 
-.contactbutton {
+/* .contactbutton {
   border: 0px;
   background-color: white;
   color: #8467da;
@@ -145,4 +145,4 @@ textarea {
   border-color: #8467da;
   color: #8467da;
   font-weight: bold;
-}
+}  */
