+++ src/components/Header.css
@@ -7,10 +7,17 @@
   align-items: center;
 }
 
-.Save {
+.settings {
   margin-right: 10px;
+  margin-top: 10px;
 }
 
-.Settings {
+/* .Settings {
   margin-left: 10rem;
-}
+} */
+
+/* @media screen and (min-width: 1100px) {
+   .Settings {
+     margin-left: 60rem;
+   }
+ } */
