+++ src/components/Header.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
 
 import Logo from '../pictures/logo.png';
-import Save from '../pictures/heart.png';
-import aboutUs from '../pictures/about-us.png';
+// import Save from '../pictures/heart.png';
+
+// import aboutUs from '../pictures/aboutUs.png';
+import settingIcon from '../pictures/Settings-icon.jpg';
 
 import './Header.css';
 
@@ -13,10 +15,7 @@ export default function Header() {
         <img src={Logo} className="Logo" height="70" width="auto" alt="logo-nasato" />{' '}
       </a>
       <a href="/Apropos">
-        <img src={aboutUs} className="Settings" height="40" width="40" color="#9050fd" alt="Save-nasato" />
-      </a>
-      <a href="/Favoris">
-        <img src={Save} className="Save" height="30" alt="Save-nasato" />
+        <img src={settingIcon} className="settings" height="60" width="60" alt="aboutUs-nasato" />
       </a>
     </div>
   );
