+++ src/components/Loader.css
@@ -7,7 +7,6 @@
 }
 
 body {
-  background-color: #ede6d8;
   overflow-y: hidden;
 }
 
@@ -21,7 +20,6 @@ main {
   height: 60px;
   perspective: 1000px;
   margin: 150px auto 0;
-  /* background-color: #ede6d8; */
   margin-top: 10rem;
   display: flex;
   justify-content: center;
@@ -41,72 +39,48 @@ main {
   height: 150px;
   border-radius: 300px;
 }
-/* .face > svg {
-  border-radius: 50px;
-} */
 
 .top {
-  /* background-color: 	#FF00FF; */
   border: 1px solid rgb(231, 227, 227);
   transform: rotateX(90deg) translateZ(100px);
   animation: balanceTop 1s ease-out forwards;
   display: flex;
   justify-content: center;
   align-items: center;
+  overflow: hidden;
 }
-/* .top > img {
-  width: 60%;
-  height: 60%;
-  border-radius: 50px;
-} */
 
 .bottom {
-  /* background-color: 	#FF00FF; */
   border: 1px solid rgb(231, 227, 227);
   transform: rotateX(-90deg) translateZ(100px);
   animation: balanceBottom 1.5s ease-out forwards;
   display: flex;
   justify-content: center;
   align-items: center;
+  overflow: hidden;
 }
-/* .bottom > img {
-  width: 60%;
-  height: 60%;
-  border-radius: 50px;
-} */
 
 .right {
-  /* background-color: 	#FF00FF; */
   border: 1px solid rgb(231, 227, 227);
   transform: rotateY(90deg) translateZ(100px);
   animation: balanceRight 2s ease-in forwards;
   display: flex;
   justify-content: center;
   align-items: center;
+  overflow: hidden;
 }
-/* .right > img {
-  width: 60%;
-  height: 60%;
-  border-radius: 50px;
-} */
 
 .left {
-  /* background-color: 	#FF00FF; */
   border: 1px solid rgb(231, 227, 227);
   transform: rotateY(-90deg) translateZ(100px);
   animation: balanceLeft 2.5s ease-in forwards;
   display: flex;
   justify-content: center;
   align-items: center;
+  overflow: hidden;
 }
-/* .left > img {
-  width: 60%;
-  height: 60%;
-  border-radius: 50px;
-} */
 
 .front {
-  /* background-color: 	#FF00FF; */
   border: 1px solid rgb(231, 227, 227);
   transform: rotateX(0deg) translateZ(100px);
   animation: balanceFront 3s ease-in forwards;
@@ -115,20 +89,15 @@ main {
   align-items: center;
   overflow: hidden;
 }
-/* .front > img {
-  width: 60%;
-  height: 60%;
-  border-radius: 50%;
-} */
 
 .back {
-  /* background-color: 	#FF00FF; */
   border: 1px solid rgb(231, 227, 227);
   transform: rotateX(-180deg) translateZ(100px);
   animation: balanceBack 3.5s ease-in forwards;
   display: flex;
   justify-content: center;
   align-items: center;
+  overflow: hidden;
 }
 
 @keyframes spin {
@@ -140,79 +109,59 @@ main {
   }
 }
 
-/* animation RIGHT */
 @keyframes balanceRight {
   0% {
     transform: rotateY(90deg) translateZ(1000px);
-    background: #ede6d8;
   }
   100% {
     transform: rotateY(90deg) translateZ(100px);
-    background-color: #ede6d8;
   }
 }
 
-/* animation LEFT */
 @keyframes balanceLeft {
   0% {
     transform: rotateY(-90deg) translateZ(1000px);
-    background: #ede6d8;
   }
   100% {
     transform: rotateY(-90deg) translateZ(100px);
-    background-color: #ede6d8;
   }
 }
 
-/* animation TOP */
 @keyframes balanceTop {
   0% {
     transform: rotateX(90deg) translateZ(1000px);
-    background: #ede6d8;
   }
   100% {
     transform: rotateX(90deg) translateZ(100px);
-    background-color: #ede6d8;
   }
 }
 
-/* animation BOTTOM */
 @keyframes balanceBottom {
   0% {
     transform: rotateX(-90deg) translateZ(1000px);
-    background: #ede6d8;
   }
   100% {
     transform: rotateX(-90deg) translateZ(100px);
-    background-color: #ede6d8;
   }
 }
 
-/* animation FRONT */
 @keyframes balanceFront {
   0% {
     transform: rotateX(0deg) translateZ(800px);
-    background: #ede6d8;
   }
   100% {
     transform: rotateX(0deg) translateZ(100px);
-    background-color: #ede6d8;
   }
 }
 
-/* animation BACK */
 @keyframes balanceBack {
   0% {
     transform: rotateX(-180deg) translateZ(800px);
-    background: #ede6d8;
   }
   100% {
     transform: rotateX(-180deg) translateZ(100px);
-    background-color: #ede6d8;
   }
 }
-
-/************   SENTENCE   *************/
 
 .sentence {
   height: 30%;
@@ -225,20 +174,52 @@ main {
 }
 .welcome {
   font-size: 50px;
+  color: #8467da;
+  font-weight: 600;
 }
 .to {
   font-size: 50px;
   margin-left: 1rem;
+  color: #8467da;
+  font-weight: 600;
 }
-.your {
+.nasato {
   font-size: 50px;
   margin-left: 1rem;
+  color: #8467da;
+  font-weight: 800;
 }
-.relaxation {
-  font-size: 50px;
-  margin-left: 1rem;
-}
-.app {
-  font-size: 50px;
-  margin-left: 1rem;
+
+@media screen and (max-width: 680px) {
+  .sentence {
+    margin-top: 11rem;
+    flex-wrap: wrap;
+    display: flex;
+    font-family: 'PT mono', sans-serif;
+    /* background: cadetblue; */
+  }
+  .container-span1 {
+    height: 20%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    /* background: burlywood; */
+  }
+  .welcome {
+    font-size: 30px;
+    color: #8467da;
+    font-weight: 600;
+  }
+  .to {
+    font-size: 30px;
+    margin-left: 0.5rem;
+    color: #8467da;
+    font-weight: 600;
+  }
+  .nasato {
+    font-size: 30px;
+    margin-left: 0.5rem;
+    color: #8467da;
+    font-weight: 800;
+  }
 }
