+++ src/components/Settings.css
@@ -1,4 +1,4 @@
-.main-container {
+/* .main-container {
   height: 100vh;
   width: 100%;
   display: flex;
@@ -39,7 +39,7 @@
 
 /********* Elisa Part *********/
 
-.elisa-container {
+/* .elisa-container {
   height: 100%;
   width: 25%;
   display: flex;
@@ -56,11 +56,11 @@
   width: 100%;
   height: 100%;
   border-radius: 50%;
-}
+} */
 
 /*********** JS PART *************/
 
-.jc-container {
+/* .jc-container {
   height: 100%;
   width: 25%;
   display: flex;
@@ -77,11 +77,11 @@
   width: 100%;
   height: 100%;
   border-radius: 50%;
-}
+} */
 
 /*********** Thomas PART *************/
 
-.thomas-container {
+/* .thomas-container {
   height: 100%;
   width: 25%;
   display: flex;
@@ -98,11 +98,11 @@
   width: 100%;
   height: 100%;
   border-radius: 50%;
-}
+} */
 
 /*********** Yassine PART *************/
 
-.yassine-container {
+/* .yassine-container {
   height: 100%;
   width: 25%;
   display: flex;
@@ -119,4 +119,4 @@
   width: 100%;
   height: 100%;
   border-radius: 50%;
-}
+} */
