+++ src/components/Settings.jsx
@@ -1,47 +1,47 @@
-import React from 'react';
-import './Settings.css';
-import Monkey from '../pictures/Monkey-2.jpg';
-import Formulaire from './Formulaire.jsx';
+// import React from 'react';
+// import './Settings.css';
+// import Monkey from '../pictures/Monkey-2.jpg';
+// import Formulaire from './Formulaire.jsx';
 
-function Settings() {
-  return (
-    <main className="main-container">
-      <p>LA PAGE SETTINGS</p>
+// function Settings() {
+//   return (
+//     <main className="main-container">
+//       <p>LA PAGE SETTINGS</p>
 
-      <div className="container-aboutUs">
-        <section className="title">
-          <h1>About Us</h1>
-        </section>
-        <div className="container">
-          <div className="elisa-container">
-            <div className="containerImg">
-              <img src={Monkey} alt="singe" height="20" width="20" />
-            </div>
-            <p>Elisa</p>
-          </div>
-          <div className="jc-container">
-            <div className="containerImg">
-              <img src={Monkey} alt="singe" height="20" width="20" />
-            </div>
-            <p>Juan Carlos</p>
-          </div>
-          <div className="thomas-container">
-            <div className="containerImg">
-              <img src={Monkey} alt="singe" height="20" width="20" />
-            </div>
-            <p>Thomas</p>
-          </div>
-          <div className="yassine-container">
-            <div className="containerImg">
-              <img src={Monkey} alt="singe" height="20" width="20" />
-            </div>
-            <p>Yassine</p>
-          </div>
-        </div>
-      </div>
-      <Formulaire />
-    </main>
-  );
-}
+//       <div className="container-aboutUs">
+//         <section className="title">
+//           <h1>About Us</h1>
+//         </section>
+//         <div className="container">
+//           <div className="elisa-container">
+//             <div className="containerImg">
+//               <img src={Monkey} alt="singe" height="20" width="20" />
+//             </div>
+//             <p>Elisa</p>
+//           </div>
+//           <div className="jc-container">
+//             <div className="containerImg">
+//               <img src={Monkey} alt="singe" height="20" width="20" />
+//             </div>
+//             <p>Juan Carlos</p>
+//           </div>
+//           <div className="thomas-container">
+//             <div className="containerImg">
+//               <img src={Monkey} alt="singe" height="20" width="20" />
+//             </div>
+//             <p>Thomas</p>
+//           </div>
+//           <div className="yassine-container">
+//             <div className="containerImg">
+//               <img src={Monkey} alt="singe" height="20" width="20" />
+//             </div>
+//             <p>Yassine</p>
+//           </div>
+//         </div>
+//       </div>
+//       <Formulaire />
+//     </main>
+//   );
+// }
 
-export default Settings;
+// export default Settings;
