+++ src/components/a-propos.jsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import './a-propos.css';
 import Team from './team.jsx';
-// import Preferences from './preferences';
+import Preferences from './preferences';
 import AboutUs from './about-us';
-import Accordion from './Accordion.jsx';
 import ContactUs from './contact-us';
 
 function Apropos() {
@@ -14,8 +13,7 @@ function Apropos() {
       </div>
       <Team />
       <AboutUs />
-      <Accordion />
-      {/* <Preferences /> */}
+      <Preferences />
       <ContactUs />
     </div>
   );
