+++ src/components/team-member.css
@@ -5,6 +5,7 @@
 .teammember {
   display: block;
   width: 100%;
+  height: 100%;
   border-radius: 50%;
   border-style: solid;
   border-color: #8467da;
